+++ changelog.md
@@ -66,6 +66,7 @@
 - The loading of `GridMaps` from files has been reworked. It should no longer be possible to load a grid map with invalid coordinates.
 - Introduced `System::guess_elements_unknown` for assigning elements only to atoms which are not already assigned an element.
 - `ProgressPrinter` now only uses mutex if the `parallel` feature is specified.
+- Introduced `Vector3D::rotate`, `Atom::rotate`, and `Atom::rotate_nopbc` for performing rotations using rotation matrices.
 - Bug fix: Regex operators can be now used in regular expression blocks without raising an InvalidOperator error.
 
 ***